Allow branding
Currently the skin is branded for draph.io (understandably!)
To make this more reusable, we can read a logo from the value of
wgLogos to allow the rendering of other logos.

The footer is also not applicable to other wikis, so I've added
a configuration variable to disable it. it defaults to true to
maintain the status quo.